fix again

+++ tasks/mpi/kalinin_d_vector_dot_product/src/ops_mpi.cpp
@@ -2,7 +2,8 @@
 #include "mpi/kalinin_d_vector_dot_product/include/ops_mpi.hpp"
 
 #include <boost/mpi/collectives.hpp>
-#include <boost/mpi/operations.hpp>
+#include <boost/mpi/collectives/broadcast.hpp>
+#include <boost/mpi/collectives/reduce.hpp>
 #include <cmath>
 #include <cstddef>
 #include <functional>
@@ -103,10 +104,10 @@ bool kalinin_d_vector_dot_product_mpi::TestMPITaskParallel::PreProcessingImpl() 
 bool kalinin_d_vector_dot_product_mpi::TestMPITaskParallel::RunImpl() {
   if (world_.rank() == 0) {
     size_t offset_remainder = counts_[0];
-    for (unsigned int proc = 1; proc < num_processes_; proc++) {
+    for (int proc = 1; proc < num_processes_; proc++) {
       size_t current_count = counts_[proc];
-      world_.send(static_cast<int>(proc), 0, input_[0].data() + offset_remainder, static_cast<int>(current_count));
-      world_.send(static_cast<int>(proc), 1, input_[1].data() + offset_remainder, static_cast<int>(current_count));
+      world_.send(proc, 0, input_[0].data() + offset_remainder, static_cast<int>(current_count));
+      world_.send(proc, 1, input_[1].data() + offset_remainder, static_cast<int>(current_count));
       offset_remainder += current_count;
     }
   }
